Refactor MTP and DFlash2
