+++ src/speculative/dflash.rs
@@ -1,12 +1,15 @@
 use crate::models::dflash::{DFlashDraftModel, DFlashModelConfig};
 use crate::models::layers::distributed::Comm;
 use crate::models::layers::VarBuilderX;
-use candle_core::{DType, Device, IndexOp, Result, Tensor, D};
+use candle_core::{DType, Device, Result, Tensor};
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::rc::Rc;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Mutex;
+
+/// DFlash2 drafts attend to a bounded projected-context window (matches reference training).
+const DEFAULT_CONTEXT_WINDOW: usize = 512;
 
 static DFLASH_TOTAL_PROPOSED: AtomicUsize = AtomicUsize::new(0);
 static DFLASH_TOTAL_ACCEPTED: AtomicUsize = AtomicUsize::new(0);
@@ -20,7 +23,7 @@ fn dflash_stats_update(proposed: usize, accepted: usize) {
         let total_proposed = DFLASH_TOTAL_PROPOSED.load(Ordering::Relaxed);
         let total_accepted = DFLASH_TOTAL_ACCEPTED.load(Ordering::Relaxed);
         crate::log_info!(
-            "DFlash Stats: steps={}, proposed={}, accepted={}, average_acceptance_rate={:.2}%",
+            "DFlash2 Stats: steps={}, proposed={}, accepted={}, average_acceptance_rate={:.2}%",
             steps,
             total_proposed,
             total_accepted,
@@ -38,16 +41,10 @@ pub struct DFlashDrafter {
     pub target_layer_ids: Vec<usize>,
     pub num_speculative_tokens: usize,
     pub mask_token_id: u32,
+    context_window: usize,
     device: Device,
     _dtype: DType,
     cached_target_hidden: Mutex<HashMap<usize, Tensor>>,
-}
-
-pub struct SpecDecodeOutput {
-    pub accepted_tokens: Vec<Vec<u32>>,
-    pub accepted_counts: Vec<usize>,
-    pub logits: Tensor,
-    pub hidden_states: Vec<Tensor>,
 }
 
 impl DFlashDrafter {
@@ -59,6 +56,13 @@ impl DFlashDrafter {
         device: &Device,
         num_speculative_tokens: Option<usize>,
     ) -> Result<Self> {
+        if !draft_config.is_dflash2() {
+            candle_core::bail!(
+                "Only DFlash2 draft models are supported (architecture DFlash2* or dflash_config.selector_top_k). \
+                 For Qwen3.5 built-in speculative decoding, use --num-speculative-tokens without --draft-model."
+            );
+        }
+
         let draft_vb = unsafe {
             candle_nn::var_builder::ShardedSafeTensors::var_builder(
                 draft_weight_files,
@@ -78,19 +82,22 @@ impl DFlashDrafter {
             DFlashDraftModel::new(&draft_vb, comm, draft_config, DType::BF16, device)?;
 
         let target_layer_ids = draft_config.target_layer_ids();
-        // DFlash config.block_size is the verification block width:
-        // [known first token] + N draft tokens. The user-facing speculative
-        // token count is N.
+        // block_size = [anchor] + N mask slots; user-facing count is N.
         let block_size =
             num_speculative_tokens.unwrap_or_else(|| draft_config.block_size().saturating_sub(1));
         let mask_token_id = draft_config.mask_token_id().unwrap_or(0);
+        let context_window = std::cmp::min(
+            DEFAULT_CONTEXT_WINDOW,
+            draft_config.max_position_embeddings.max(1),
+        );
 
         crate::log_info!(
-            "DFlash drafter initialized: {} layers, num_speculative_tokens={}, target_layer_ids={:?}, mask_token_id={}",
+            "DFlash2 drafter initialized: {} layers, num_speculative_tokens={}, target_layer_ids={:?}, mask_token_id={}, context_window={}",
             draft_config.num_hidden_layers,
             block_size,
             target_layer_ids,
             mask_token_id,
+            context_window,
         );
 
         Ok(Self {
@@ -98,6 +105,7 @@ impl DFlashDrafter {
             target_layer_ids,
             num_speculative_tokens: block_size,
             mask_token_id,
+            context_window,
             device: device.clone(),
             _dtype: dtype,
             cached_target_hidden: Mutex::new(HashMap::new()),
@@ -113,24 +121,17 @@ impl DFlashDrafter {
         self.draft_model.project_layer_hiddens(layer_hiddens)
     }
 
+    /// Run the DFlash2 draft block and select tokens via the top-k candidate lattice.
     pub fn draft_tokens(
         &self,
         target_hidden: &Tensor,
         embed_fn: &dyn Fn(&Tensor) -> Result<Tensor>,
         lm_head_fn: &dyn Fn(&Tensor) -> Result<Tensor>,
-        last_tokens: &[u32],
+        anchor_token: u32,
     ) -> Result<Vec<u32>> {
-        let batch_size = last_tokens.len();
-        assert_eq!(
-            batch_size, 1,
-            "DFlash currently supports batch_size=1 for drafting"
-        );
-
         let n = self.num_speculative_tokens;
-        let mut draft_token_ids: Vec<u32> = Vec::with_capacity(n);
-
         let mut block_ids = vec![self.mask_token_id; n + 1];
-        block_ids[0] = last_tokens[0];
+        block_ids[0] = anchor_token;
 
         let block_tensor = Tensor::from_vec(
             block_ids.iter().map(|&x| x as i64).collect::<Vec<_>>(),
@@ -138,8 +139,7 @@ impl DFlashDrafter {
             &self.device,
         )?;
 
-        let noise_embedding = embed_fn(&block_tensor)?;
-        let noise_embedding = noise_embedding.to_dtype(DType::BF16)?;
+        let noise_embedding = embed_fn(&block_tensor)?.to_dtype(DType::BF16)?;
 
         let target_hidden_2d = if target_hidden.rank() == 3 {
             let (_, ctx, h) = target_hidden.dims3()?;
@@ -164,69 +164,11 @@ impl DFlashDrafter {
         let draft_hidden =
             self.draft_model
                 .forward(&target_hidden_bf16, &noise_2d, &positions_tensor)?;
-
         let total_out = draft_hidden.dim(0)?;
         let draft_hidden = draft_hidden.narrow(0, total_out - n, n)?;
         let draft_logits = lm_head_fn(&draft_hidden)?;
-        // DFlash2: candidate selector walks a top-k lattice. Keep greedy argmax as
-        // fallback when selector is unavailable.
-        if self.draft_model.is_dflash2() {
-            return self.draft_model.select_candidates(
-                &draft_hidden,
-                &draft_logits,
-                last_tokens[0],
-            );
-        }
-
-        for i in 0..n {
-            let logit_slice = draft_logits.i(i)?;
-            let argmax_result = logit_slice.argmax(D::Minus1)?;
-            let token_id = if argmax_result.rank() > 0 {
-                argmax_result.flatten_all()?.i(0)?.to_vec0::<u32>()?
-            } else {
-                argmax_result.to_vec0::<u32>()?
-            };
-            draft_token_ids.push(token_id);
-        }
-
-        Ok(draft_token_ids)
-    }
-
-    pub fn verify_tokens(
-        draft_tokens: &[u32],
-        target_logits: &Tensor,
-        _temperature: f32,
-    ) -> Result<(Vec<u32>, usize)> {
-        let n_draft = draft_tokens.len();
-        let mut accepted = Vec::new();
-
-        for i in 0..n_draft {
-            let argmax_res = target_logits.i(i)?.argmax(D::Minus1)?;
-            let target_token = if argmax_res.rank() > 0 {
-                argmax_res.flatten_all()?.i(0)?.to_vec0::<u32>()?
-            } else {
-                argmax_res.to_vec0::<u32>()?
-            };
-
-            if i < n_draft && target_token == draft_tokens[i] {
-                accepted.push(target_token);
-            } else {
-                accepted.push(target_token);
-                let len = accepted.len();
-                return Ok((accepted, len));
-            }
-        }
-
-        let bonus_argmax = target_logits.i(n_draft)?.argmax(D::Minus1)?;
-        let bonus_token = if bonus_argmax.rank() > 0 {
-            bonus_argmax.flatten_all()?.i(0)?.to_vec0::<u32>()?
-        } else {
-            bonus_argmax.to_vec0::<u32>()?
-        };
-        accepted.push(bonus_token);
-
-        let len = accepted.len();
-        Ok((accepted, len))
+        self.draft_model
+            .select_candidates(&draft_hidden, &draft_logits, anchor_token)
     }
 
     pub fn target_layer_ids(&self) -> &[usize] {
@@ -241,62 +183,93 @@ impl DFlashDrafter {
         self.cached_target_hidden.lock().unwrap().remove(&seq_id);
     }
 
-    /// Return the accumulated hidden state context for drafting, or None if empty.
     pub fn build_draft_context(&self, seq_id: usize) -> Result<Option<Tensor>> {
         let cached = self.cached_target_hidden.lock().unwrap();
         Ok(cached.get(&seq_id).cloned())
     }
 
-    /// After verification, append hidden states for the verified input tokens.
-    /// verify_hidden covers [first_token, d0, ..., d_{n-1}] at rows [0, 1, ..., n].
-    /// The recovered/bonus token is produced by logits and is not part of this forward input,
-    /// so the correct number of rows to keep is accepted_count.
-    pub fn replace_with_verified_hidden(
+    pub fn append_context(&self, seq_id: usize, projected: &Tensor) -> Result<()> {
+        let rows = projected.dim(0)?;
+        if rows == 0 {
+            return Ok(());
+        }
+        let mut cached = self.cached_target_hidden.lock().unwrap();
+        let updated = match cached.get(&seq_id).cloned() {
+            Some(prev) => Tensor::cat(&[prev, projected.clone()], 0)?,
+            None => projected.clone(),
+        };
+        let total = updated.dim(0)?;
+        let keep = std::cmp::min(total, self.context_window);
+        cached.insert(seq_id, updated.narrow(0, total - keep, keep)?);
+        Ok(())
+    }
+
+    pub fn append_verified_context(
         &self,
         verify_hidden: &Tensor,
         accepted_count: usize,
         seq_id: usize,
     ) -> Result<()> {
-        let mut cached = self.cached_target_hidden.lock().unwrap();
-
         let vdim = verify_hidden.dim(0)?;
         if accepted_count == 0 || vdim == 0 {
             return Ok(());
         }
-
-        let keep = std::cmp::min(accepted_count, vdim);
-        let verified_inputs = verify_hidden.narrow(0, 0, keep)?;
-
-        if let Some(prev) = cached.get(&seq_id).cloned() {
-            cached.insert(seq_id, Tensor::cat(&[prev, verified_inputs], 0)?);
-        } else {
-            cached.insert(seq_id, verified_inputs);
-        }
-        Ok(())
+        let keep = std::cmp::min(accepted_count + 1, vdim);
+        self.append_context(seq_id, &verify_hidden.narrow(0, 0, keep)?)
     }
 
-    /// Store hidden states from a forward pass (prefill or decode).
-    /// Appends to the accumulated context.
     pub fn store_decode_hidden(&self, hidden: &Tensor, seq_id: usize) -> Result<()> {
-        let mut cached = self.cached_target_hidden.lock().unwrap();
-        if let Some(prev) = cached.get(&seq_id).cloned() {
-            cached.insert(seq_id, Tensor::cat(&[prev, hidden.clone()], 0)?);
-        } else {
-            cached.insert(seq_id, hidden.clone());
-        }
-        Ok(())
+        self.append_context(seq_id, hidden)
     }
 }
 
-use crate::core::mtp::{verify_draft_greedy, MtpSeqInfo};
 use crate::core::runner::{Model, ModelRunner, Seqs};
 use crate::models::layers::linear::set_linear_is_prefill;
+use crate::speculative::metadata::SpecSeqInfo;
+use crate::speculative::verify::verify_draft_greedy;
+use crate::utils::config::EngineConfig;
+use attention_rs::InputMetadata;
+
+/// Collect target-layer hidden states for DFlash2 context / verify refresh.
+fn forward_collecting_target(
+    model: &Model,
+    input_ids: &Tensor,
+    positions: &Tensor,
+    kv_pairs: Option<&Vec<(Tensor, Tensor)>>,
+    input_metadata: &InputMetadata,
+    target_layer_ids: &[usize],
+) -> Result<(Tensor, Vec<Tensor>)> {
+    match model {
+        Model::Qwen3_5(m) => m.forward_collecting_layers(
+            input_ids,
+            positions,
+            kv_pairs,
+            input_metadata,
+            false,
+            target_layer_ids,
+        ),
+        Model::Qwen3_5MoE(m) => m.forward_with_hidden_states(
+            input_ids,
+            positions,
+            kv_pairs,
+            input_metadata,
+            false,
+            target_layer_ids,
+        ),
+        Model::Qwen3VL(m) => m.forward_with_hidden_states(
+            input_ids,
+            positions,
+            kv_pairs,
+            input_metadata,
+            false,
+            target_layer_ids,
+        ),
+        _ => candle_core::bail!("DFlash2 supports Qwen3.5 / Qwen3.5 MoE / Qwen3-VL targets"),
+    }
+}
 
 impl ModelRunner {
-    /// Run one external DFlash speculative step.
-    ///
-    /// DFlash shares the MTP verification metadata and Mamba rollback path,
-    /// while its draft model consumes projected intermediate target states.
+    /// One DFlash2 speculative decode step (anchor + draft + verify).
     pub fn run_dflash_decode(&self, seqs: Seqs) -> Result<Vec<Vec<u32>>> {
         let batch_size = match seqs {
             Seqs::SeqRefs(ref refs) => refs.len(),
@@ -311,12 +284,12 @@ impl ModelRunner {
         };
 
         let seq_info = match seqs {
-            Seqs::SeqRefs(ref refs) if refs.len() == 1 => MtpSeqInfo {
+            Seqs::SeqRefs(ref refs) if refs.len() == 1 => SpecSeqInfo {
                 id: refs[0].id,
                 len: refs[0].len(),
                 block_table: refs[0].block_table.clone(),
             },
-            Seqs::DecodeVec(ref decoded) if decoded.len() == 1 => MtpSeqInfo {
+            Seqs::DecodeVec(ref decoded) if decoded.len() == 1 => SpecSeqInfo {
                 id: decoded[0].id,
                 len: decoded[0].len,
                 block_table: decoded[0].block_tables.clone(),
@@ -357,33 +330,14 @@ impl ModelRunner {
         let (anchor_logits, anchor_hidden_states) = {
             let kv_cache = self.get_kv_cache();
             let kv_pairs = kv_cache.as_pairs();
-            let result = match self.model() {
-                Model::Qwen3_5(model) => model.forward_with_hidden_states(
-                    &input_ids,
-                    &positions,
-                    kv_pairs,
-                    &input_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                Model::Qwen3_5MoE(model) => model.forward_with_hidden_states(
-                    &input_ids,
-                    &positions,
-                    kv_pairs,
-                    &input_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                Model::Qwen3VL(model) => model.forward_with_hidden_states(
-                    &input_ids,
-                    &positions,
-                    kv_pairs,
-                    &input_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
-            }?;
+            let result = forward_collecting_target(
+                self.model(),
+                &input_ids,
+                &positions,
+                kv_pairs,
+                &input_metadata,
+                drafter.target_layer_ids(),
+            )?;
             drop(kv_cache);
             result
         };
@@ -395,18 +349,18 @@ impl ModelRunner {
             .as_ref()
             .map_or(true, |context| context.dim(0).unwrap_or(0) < seq_info.len)
         {
-            drafter.store_decode_hidden(&projected_step_hidden, seq_info.id)?;
+            drafter.append_context(seq_info.id, &projected_step_hidden)?;
         }
         let target_hidden = drafter
             .build_draft_context(seq_info.id)?
-            .ok_or_else(|| candle_core::Error::Msg("DFlash target hidden cache is empty".into()))?;
+            .ok_or_else(|| candle_core::Error::Msg("DFlash2 context cache is empty".into()))?;
 
         let embed_fn = |tokens: &Tensor| -> Result<Tensor> {
             match self.model() {
                 Model::Qwen3_5(model) => model.embed_forward(tokens),
                 Model::Qwen3_5MoE(model) => model.embed_forward(tokens),
                 Model::Qwen3VL(model) => model.embed_forward(tokens),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
+                _ => candle_core::bail!("DFlash2 supports Qwen3.5 family targets"),
             }
         };
         let lm_head_fn = |hidden: &Tensor| -> Result<Tensor> {
@@ -414,11 +368,11 @@ impl ModelRunner {
                 Model::Qwen3_5(model) => model.forward_lm_head(hidden),
                 Model::Qwen3_5MoE(model) => model.forward_lm_head(hidden),
                 Model::Qwen3VL(model) => model.forward_lm_head(hidden),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
+                _ => candle_core::bail!("DFlash2 supports Qwen3.5 family targets"),
             }
         };
         let draft_tokens =
-            drafter.draft_tokens(&target_hidden, &embed_fn, &lm_head_fn, &[anchor_token])?;
+            drafter.draft_tokens(&target_hidden, &embed_fn, &lm_head_fn, anchor_token)?;
         if draft_tokens.is_empty() {
             return Ok(vec![vec![anchor_token]]);
         }
@@ -427,7 +381,7 @@ impl ModelRunner {
         verify_tokens.extend_from_slice(&draft_tokens);
         let verify_len = verify_tokens.len();
         let slot_mappings =
-            self.compute_slot_mappings(&seq_info, verify_len, self.block_size(), "DFlash verify")?;
+            self.compute_slot_mappings(&seq_info, verify_len, self.block_size(), "DFlash2 verify")?;
         let verify_input_ids = Tensor::from_vec(verify_tokens, (verify_len,), self.device())?;
         let verify_positions = Tensor::from_vec(
             (seq_info.len..seq_info.len + verify_len)
@@ -436,12 +390,9 @@ impl ModelRunner {
             (verify_len,),
             self.device(),
         )?;
-        let verify_metadata = self.build_mtp_metadata(&seq_info, &slot_mappings, verify_len)?;
+        let verify_metadata = self.build_verify_metadata(&seq_info, &slot_mappings, verify_len)?;
         let _verify_guard = set_linear_is_prefill(true);
 
-        // MTP-style: use CUDA-graph verify when captured. build_mtp_metadata sets
-        // flashinfer use_cuda_graph / prefill_plan_info based on the same condition,
-        // so eager and graph paths must stay in sync.
         #[cfg(all(feature = "cuda", feature = "graph"))]
         let use_verify_graph = self
             .mtp_capturer
@@ -453,8 +404,6 @@ impl ModelRunner {
         let (verify_logits, projected_verify_hidden) = if use_verify_graph {
             #[cfg(all(feature = "cuda", feature = "graph"))]
             {
-                // Hold mamba cache across verify-graph replay (same as decode graphs):
-                // GDN gather/scatter mutates cache storage in-place under the graph.
                 let logits = match self.model() {
                     Model::Qwen3_5(model) => {
                         let _guard = model.lock_mamba_cache_for_graph();
@@ -501,7 +450,7 @@ impl ModelRunner {
                 }
                 .ok_or_else(|| {
                     candle_core::Error::Msg(
-                        "DFlash verify missing layer-hidden buffers after graph replay".into(),
+                        "DFlash2 verify missing layer-hidden buffers after graph replay".into(),
                     )
                 })?;
                 (logits, drafter.project_layer_hiddens(&layer_hiddens)?)
@@ -513,49 +462,37 @@ impl ModelRunner {
         } else {
             let kv_cache = self.get_kv_cache();
             let kv_pairs = kv_cache.as_pairs();
-            let (logits, hidden_states) = match self.model() {
-                Model::Qwen3_5(model) => model.forward_with_hidden_states(
-                    &verify_input_ids,
-                    &verify_positions,
-                    kv_pairs,
-                    &verify_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                )?,
-                Model::Qwen3_5MoE(model) => model.forward_with_hidden_states(
-                    &verify_input_ids,
-                    &verify_positions,
-                    kv_pairs,
-                    &verify_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                )?,
-                Model::Qwen3VL(model) => model.forward_with_hidden_states(
-                    &verify_input_ids,
-                    &verify_positions,
-                    kv_pairs,
-                    &verify_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                )?,
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
-            };
+            let (verify_logits, hidden_states) = forward_collecting_target(
+                self.model(),
+                &verify_input_ids,
+                &verify_positions,
+                kv_pairs,
+                &verify_metadata,
+                drafter.target_layer_ids(),
+            )?;
             drop(kv_cache);
-            (logits, drafter.extract_and_concat_hidden(&hidden_states)?)
+            (
+                verify_logits,
+                drafter.extract_and_concat_hidden(&hidden_states)?,
+            )
         };
 
         let verify_result = verify_draft_greedy(&verify_logits, &draft_tokens)?;
         let commit_len = 1 + verify_result.num_accepted;
-        if verify_result.num_accepted < verify_result.num_proposed {
-            if !self.mtp_rollback_mamba(seq_info.id, commit_len)? {
-                candle_core::bail!(
-                    "DFlash failed to roll back mamba state for sequence {}",
-                    seq_info.id
-                );
-            }
+        if verify_result.num_accepted < verify_result.num_proposed
+            && !self.rollback_mamba(seq_info.id, commit_len)?
+        {
+            candle_core::bail!(
+                "DFlash2 failed to roll back mamba state for sequence {}",
+                seq_info.id
+            );
         }
 
-        drafter.replace_with_verified_hidden(&projected_verify_hidden, commit_len, seq_info.id)?;
+        drafter.append_verified_context(
+            &projected_verify_hidden,
+            verify_result.num_accepted,
+            seq_info.id,
+        )?;
 
         let mut result_tokens = Vec::with_capacity(commit_len + 1);
         result_tokens.push(anchor_token);
@@ -573,7 +510,7 @@ impl ModelRunner {
         let seq_infos = match seqs {
             Seqs::SeqRefs(ref refs) => refs
                 .iter()
-                .map(|seq| MtpSeqInfo {
+                .map(|seq| SpecSeqInfo {
                     id: seq.id,
                     len: seq.len(),
                     block_table: seq.block_table.clone(),
@@ -581,7 +518,7 @@ impl ModelRunner {
                 .collect::<Vec<_>>(),
             Seqs::DecodeVec(ref decoded) => decoded
                 .iter()
-                .map(|seq| MtpSeqInfo {
+                .map(|seq| SpecSeqInfo {
                     id: seq.id,
                     len: seq.len,
                     block_table: seq.block_tables.clone(),
@@ -618,48 +555,21 @@ impl ModelRunner {
         let (anchor_logits, anchor_hidden_states) = {
             let kv_cache = self.get_kv_cache();
             let kv_pairs = kv_cache.as_pairs();
-            let result = match self.model() {
-                Model::Qwen3_5(model) => model.forward_with_hidden_states(
-                    &input_ids,
-                    &positions,
-                    kv_pairs,
-                    &input_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                Model::Qwen3_5MoE(model) => model.forward_with_hidden_states(
-                    &input_ids,
-                    &positions,
-                    kv_pairs,
-                    &input_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                Model::Qwen3VL(model) => model.forward_with_hidden_states(
-                    &input_ids,
-                    &positions,
-                    kv_pairs,
-                    &input_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
-            }?;
+            let result = forward_collecting_target(
+                self.model(),
+                &input_ids,
+                &positions,
+                kv_pairs,
+                &input_metadata,
+                drafter.target_layer_ids(),
+            )?;
             drop(kv_cache);
             result
         };
         let anchor_tokens = self.sample(&anchor_logits, seqs, false)?;
         let projected_step_hidden = drafter.extract_and_concat_hidden(&anchor_hidden_states)?;
         for (index, seq_info) in seq_infos.iter().enumerate() {
-            if drafter
-                .build_draft_context(seq_info.id)?
-                .map_or(true, |context| context.dim(0).unwrap_or(0) < seq_info.len)
-            {
-                drafter.store_decode_hidden(
-                    &projected_step_hidden.narrow(0, index, 1)?,
-                    seq_info.id,
-                )?;
-            }
+            drafter.append_context(seq_info.id, &projected_step_hidden.narrow(0, index, 1)?)?;
         }
 
         let embed_fn = |tokens: &Tensor| -> Result<Tensor> {
@@ -667,7 +577,7 @@ impl ModelRunner {
                 Model::Qwen3_5(model) => model.embed_forward(tokens),
                 Model::Qwen3_5MoE(model) => model.embed_forward(tokens),
                 Model::Qwen3VL(model) => model.embed_forward(tokens),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
+                _ => candle_core::bail!("DFlash2 supports Qwen3.5 family targets"),
             }
         };
         let lm_head_fn = |hidden: &Tensor| -> Result<Tensor> {
@@ -675,20 +585,20 @@ impl ModelRunner {
                 Model::Qwen3_5(model) => model.forward_lm_head(hidden),
                 Model::Qwen3_5MoE(model) => model.forward_lm_head(hidden),
                 Model::Qwen3VL(model) => model.forward_lm_head(hidden),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
+                _ => candle_core::bail!("DFlash2 supports Qwen3.5 family targets"),
             }
         };
 
         let mut draft_tokens = Vec::with_capacity(seq_infos.len());
         for (seq_info, &anchor_token) in seq_infos.iter().zip(&anchor_tokens) {
-            let target_hidden = drafter.build_draft_context(seq_info.id)?.ok_or_else(|| {
-                candle_core::Error::Msg("DFlash target hidden cache is empty".into())
-            })?;
+            let target_hidden = drafter
+                .build_draft_context(seq_info.id)?
+                .ok_or_else(|| candle_core::Error::Msg("DFlash2 context cache is empty".into()))?;
             draft_tokens.push(drafter.draft_tokens(
                 &target_hidden,
                 &embed_fn,
                 &lm_head_fn,
-                &[anchor_token],
+                anchor_token,
             )?);
         }
         let verify_len = 1 + drafter.num_speculative_tokens;
@@ -711,7 +621,7 @@ impl ModelRunner {
                     seq_info,
                     verify_len,
                     self.block_size(),
-                    "DFlash batch verify",
+                    "DFlash2 batch verify",
                 )
             })
             .collect::<Result<Vec<_>>>()?;
@@ -730,40 +640,20 @@ impl ModelRunner {
             (seq_infos.len() * verify_len,),
             self.device(),
         )?;
-        let verify_metadata = self.build_mtp_metadata_batch(&seq_infos, &slot_mappings, &q_lens)?;
+        let verify_metadata =
+            self.build_verify_metadata_batch(&seq_infos, &slot_mappings, &q_lens)?;
         let _verify_guard = set_linear_is_prefill(true);
-        // Batch verify stays on the same forward_with_hidden_states path; single-seq
-        // uses the MTP-style CUDA-graphable forward + layer buffers above.
         let (verify_logits, verify_hidden_states) = {
             let kv_cache = self.get_kv_cache();
             let kv_pairs = kv_cache.as_pairs();
-            let result = match self.model() {
-                Model::Qwen3_5(model) => model.forward_with_hidden_states(
-                    &verify_input_ids,
-                    &verify_positions,
-                    kv_pairs,
-                    &verify_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                Model::Qwen3_5MoE(model) => model.forward_with_hidden_states(
-                    &verify_input_ids,
-                    &verify_positions,
-                    kv_pairs,
-                    &verify_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                Model::Qwen3VL(model) => model.forward_with_hidden_states(
-                    &verify_input_ids,
-                    &verify_positions,
-                    kv_pairs,
-                    &verify_metadata,
-                    false,
-                    drafter.target_layer_ids(),
-                ),
-                _ => candle_core::bail!("DFlash currently supports Qwen3.5 target models"),
-            }?;
+            let result = forward_collecting_target(
+                self.model(),
+                &verify_input_ids,
+                &verify_positions,
+                kv_pairs,
+                &verify_metadata,
+                drafter.target_layer_ids(),
+            )?;
             drop(kv_cache);
             result
         };
@@ -780,16 +670,16 @@ impl ModelRunner {
             let verify_result = verify_draft_greedy(&per_seq_logits, drafts)?;
             let commit_len = 1 + verify_result.num_accepted;
             if verify_result.num_accepted < verify_result.num_proposed
-                && !self.mtp_rollback_mamba_at(seq_info.id, commit_len, offset)?
+                && !self.rollback_mamba_at(seq_info.id, commit_len, offset)?
             {
                 candle_core::bail!(
-                    "DFlash failed to roll back mamba state for sequence {}",
+                    "DFlash2 failed to roll back mamba state for sequence {}",
                     seq_info.id
                 );
             }
-            drafter.replace_with_verified_hidden(
+            drafter.append_verified_context(
                 &projected_verify_hidden.narrow(0, offset, verify_len)?,
-                commit_len,
+                verify_result.num_accepted,
                 seq_info.id,
             )?;
             let mut output = Vec::with_capacity(commit_len + 1);
@@ -801,4 +691,41 @@ impl ModelRunner {
         }
         Ok(result)
     }
+}
+
+pub fn init_dflash_drafter(
+    econfig: &EngineConfig,
+    comm: Rc<Comm>,
+    device: &Device,
+) -> Result<Option<DFlashDrafter>> {
+    let Some(draft_model) = econfig.draft_model.as_ref() else {
+        return Ok(None);
+    };
+    if draft_model.is_empty() {
+        return Ok(None);
+    }
+
+    crate::log_info!("Loading external DFlash2 draft model...");
+    let (model_id, weight_path) = crate::speculative::resolve_draft_model(draft_model);
+    let loader = crate::utils::downloader::Downloader::new(model_id, weight_path, None);
+    let (draft_paths, is_gguf) = loader
+        .prepare_draft_model_weights(econfig.hf_token.clone(), econfig.hf_token_path.clone())?;
+    if is_gguf {
+        candle_core::bail!("DFlash2 draft models must use safetensors weights");
+    }
+
+    let config_data = std::fs::read(draft_paths.get_config_filename())
+        .map_err(|e| candle_core::Error::Msg(format!("Failed to read DFlash2 config: {e}")))?;
+    let draft_config: DFlashModelConfig = serde_json::from_slice(&config_data)
+        .map_err(|e| candle_core::Error::Msg(format!("Failed to parse DFlash2 config: {e}")))?;
+    let drafter = DFlashDrafter::new(
+        &draft_config,
+        &draft_paths.get_weight_filenames(),
+        comm,
+        DType::BF16,
+        device,
+        econfig.num_speculative_tokens,
+    )?;
+    crate::log_info!("External DFlash2 draft model loaded successfully");
+    Ok(Some(drafter))
 }
