+++ src/speculative/verify.rs
@@ -1,0 +1,122 @@
+use candle_core::{Result, Tensor, D};
+use std::sync::atomic::{AtomicUsize, Ordering};
+
+/// Outcome of speculative verification for a single sequence.
+#[derive(Debug, Clone)]
+pub struct MtpVerifyResult {
+    pub accepted_tokens: Vec<u32>,
+    pub continuation_token: u32,
+    pub num_accepted: usize,
+    pub num_proposed: usize,
+}
+
+/// Verify draft tokens against target model logits (greedy / argmax).
+pub fn verify_draft_greedy(
+    verify_logits: &Tensor,
+    draft_tokens: &[u32],
+) -> Result<MtpVerifyResult> {
+    let num_positions = verify_logits.dim(0)?;
+    let num_proposed = draft_tokens.len();
+
+    if num_positions == 0 || num_proposed == 0 {
+        let first_token = if num_positions > 0 {
+            verify_logits
+                .get(0)?
+                .argmax(D::Minus1)?
+                .to_scalar::<u32>()?
+        } else {
+            0
+        };
+        return Ok(MtpVerifyResult {
+            accepted_tokens: vec![],
+            continuation_token: first_token,
+            num_accepted: 0,
+            num_proposed,
+        });
+    }
+
+    let verify_logits = verify_logits.to_dtype(candle_core::DType::F32)?;
+    let all_target_tokens = verify_logits.argmax(D::Minus1)?;
+    let target_vec: Vec<u32> = all_target_tokens.to_vec1()?;
+
+    let compare_len = num_proposed.min(num_positions);
+    let mut num_accepted = 0;
+    for i in 0..compare_len {
+        if target_vec[i] == draft_tokens[i] {
+            num_accepted += 1;
+        } else {
+            break;
+        }
+    }
+
+    let accepted_tokens = draft_tokens[..num_accepted].to_vec();
+    let continuation_token = if num_accepted < num_positions {
+        target_vec[num_accepted]
+    } else {
+        target_vec[num_positions - 1]
+    };
+
+    Ok(MtpVerifyResult {
+        accepted_tokens,
+        continuation_token,
+        num_accepted,
+        num_proposed,
+    })
+}
+
+pub static MTP_TOTAL_PROPOSED: AtomicUsize = AtomicUsize::new(0);
+pub static MTP_TOTAL_ACCEPTED: AtomicUsize = AtomicUsize::new(0);
+pub static MTP_TOTAL_STEPS: AtomicUsize = AtomicUsize::new(0);
+
+pub fn mtp_stats_update(proposed: usize, accepted: usize) {
+    MTP_TOTAL_PROPOSED.fetch_add(proposed, Ordering::Relaxed);
+    MTP_TOTAL_ACCEPTED.fetch_add(accepted, Ordering::Relaxed);
+    MTP_TOTAL_STEPS.fetch_add(1, Ordering::Relaxed);
+}
+
+pub fn mtp_stats_acceptance_rate() -> f64 {
+    let proposed = MTP_TOTAL_PROPOSED.load(Ordering::Relaxed);
+    let accepted = MTP_TOTAL_ACCEPTED.load(Ordering::Relaxed);
+    if proposed == 0 {
+        0.0
+    } else {
+        accepted as f64 / proposed as f64
+    }
+}
+
+pub fn mtp_stats_avg_tokens_per_step() -> f64 {
+    let steps = MTP_TOTAL_STEPS.load(Ordering::Relaxed);
+    let accepted = MTP_TOTAL_ACCEPTED.load(Ordering::Relaxed);
+    if steps == 0 {
+        1.0
+    } else {
+        (accepted + 2 * steps) as f64 / steps as f64
+    }
+}
+
+pub fn mtp_stats_summary() -> String {
+    let proposed = MTP_TOTAL_PROPOSED.load(Ordering::Relaxed);
+    let accepted = MTP_TOTAL_ACCEPTED.load(Ordering::Relaxed);
+    let steps = MTP_TOTAL_STEPS.load(Ordering::Relaxed);
+    format!(
+        "MTP Stats: proposed={}, accepted={}, acceptance_rate={:.2}%, avg_tokens/step={:.2}",
+        proposed,
+        accepted,
+        if proposed > 0 {
+            accepted as f64 / proposed as f64 * 100.0
+        } else {
+            0.0
+        },
+        if steps > 0 {
+            (accepted + 2 * steps) as f64 / steps as f64
+        } else {
+            1.0
+        },
+    )
+}
+
+pub fn mtp_stats_reset() {
+    MTP_TOTAL_PROPOSED.store(0, Ordering::Relaxed);
+    MTP_TOTAL_ACCEPTED.store(0, Ordering::Relaxed);
+    MTP_TOTAL_STEPS.store(0, Ordering::Relaxed);
+}
